uses waitpid on relibc

+++ elysiaos-syslib/src/syscalls.rs
@@ -4,8 +4,9 @@ pub mod filesystem;
 pub mod futex;
 pub mod object;
 
-pub fn wait_for_process_exit(process: u64, exit_code_ptr: *mut u64) -> SyscallResult {
-    syscall!(WaitForProcessExit, process, exit_code_ptr as u64)
+// Wait for a process to exit
+pub fn wait_for_process_exit(process: i32, exit_code_ptr: *mut i32) -> SyscallResult {
+    syscall!(WaitForProcessExit, process as u64, exit_code_ptr as u64)
 }
 
 pub fn exit(code: u64) -> SyscallResult {
